bug fix for stats page
shows int values, not decimals. added C0BU5 script to enemy script folder (empty). Added C0BU5 Script to C0BU5 prefab

+++ surrounded/Assets/Scripts/Player/PlayerController.cs
@@ -108,7 +108,7 @@ public class PlayerController : MonoBehaviour
         XP += levelReq;
     }
     public void Stats(){
-        stats.text = "ATK: " + damage + " DEF: " + defense + " SPD: " + moveSpeed;
-        sstats.text = "ATK: " + damage + " DEF: " + defense + " SPD: " + moveSpeed;
+        stats.text = "ATK: " + damage + " DEF: " + defense + " SPD: " + (int)moveSpeed;
+        sstats.text = "ATK: " + damage + " DEF: " + defense + " SPD: " + (int)moveSpeed;
     }
 }
